source/requirements.txt
flask
google-cloud-speech
google-cloud-texttospeech
openai
janome
pyaudio
requests
natto-py
python-dotenv

